scripts/spectract/requirements.txt
"camelot-py[cv]"
pandas
pyyaml
